backend/requirements.txt
# FastAPI Core
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Environment & Configuration
python-dotenv==1.0.0

# Data Validation
pydantic[email]==2.5.3
pydantic-settings==2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# File Upload
python-multipart==0.0.6

# Image Processing & QR Codes
Pillow==10.2.0
qrcode[pil]==7.4.2

# PDF Generation
reportlab==4.0.9

# Email Templates
jinja2==3.1.3

# Email Sending
aiosmtplib==3.0.1

# Additional Utilities
python-dateutil==2.8.2